backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.13
langchain-openai==0.1.0
langchain-anthropic==0.1.4
langchain-experimental==0.0.55
pandas==2.2.1
python-multipart==0.0.9
sqlalchemy==1.4.46
fpdf2==2.7.8
matplotlib==3.8.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pymongo==4.6.3
snowflake-sqlalchemy==1.5.1
